Refactor - Option 선택 로직 분리
