+++ components/items/index.ts
@@ -1,3 +1,4 @@
 export { default as TextsSection } from './TextsSection/TextsSection';
 export { default as OptionSelector } from './OptionSelector/OptionSelector';
 export { default as SelectedOptionViewer } from './SelectedOptionViewer/SelectedOptionViewer';
+export { default as OptionViewer } from './OptionViewer/OptionViewer';
