+++ pages/items/[conItemId].tsx
@@ -1,9 +1,8 @@
-import React, { ReactNode, useState } from 'react';
+import React, { ReactNode } from 'react';
 import { AppLayout } from 'components/common';
-import { TextsSection, OptionSelector, SelectedOptionViewer } from 'components/items';
+import { OptionViewer, TextsSection } from 'components/items';
 import * as S from 'components/items/Items.styled';
 import { get } from 'apis/requestAPIs/items';
-import { commaNumber, dateFormatting } from 'utils';
 import { useRouter } from 'next/router';
 import useSWR from 'swr';
 
@@ -13,14 +12,10 @@ const Items = () => {
   const router = useRouter();
   const { conItemId } = router.query;
 
-  const { data, error } = useSWR(conItemId, (conItemId: string) => get.items(conItemId));
+  const { data } = useSWR(conItemId, (conItemId: string) => get.items(conItemId));
 
-  const [isActivePurchaseBtn, setIsActivePurchaseBtn] = useState(true);
-  const [isActiveOptionSelector, setIsActiveOptionSelector] = useState(false);
-  const [selectedOptionId, setSelectedOptionId] = useState<number>(-1);
-
-  if (!data || error) {
-    return <div>에러 발생</div>;
+  if (!data) {
+    return <div>loadding</div>;
   }
 
   const {
@@ -45,43 +40,6 @@ const Items = () => {
     ?.slice(1)
     ?.map(indexGenerator);
 
-  const optionsWithDiscountRate = options.map((option, index) => ({
-    ...option,
-    id: index,
-    discountRate: Math.floor((1 - minSellingPrice / originalPrice) * 100),
-  }));
-
-  const handleSelector = (state: 'open' | 'close') => {
-    if (!isActivePurchaseBtn && state === 'open') return;
-    setSelectedOptionId(-1);
-    setIsActivePurchaseBtn(state === 'close');
-    setIsActiveOptionSelector(state === 'open');
-  };
-
-  const handlePurchaseBtnClick = () => {
-    if (!isActivePurchaseBtn && isActiveOptionSelector && selectedOptionId === -1) {
-      return;
-    }
-    if (selectedOptionId !== -1) {
-      window.alert('베타 버전에서는 구매하실 수 없습니다.');
-      return;
-    }
-    handleSelector('open');
-  };
-
-  const handleOptionSelectorClick = (optionId: number) => {
-    setSelectedOptionId(optionId);
-    setIsActiveOptionSelector(false);
-    setIsActivePurchaseBtn(true);
-  };
-
-  const getSelectedOptionText = (selectedOptionId: number) => {
-    const option = optionsWithDiscountRate.find(({ id }) => id === selectedOptionId);
-    return typeof option?.expireAt === 'string'
-      ? `${dateFormatting(option.expireAt)} 까지 / ${commaNumber(option.sellingPrice)}원`
-      : '선택된 옵션이 없습니다.';
-  };
-
   return (
     <S.ItemsLayout>
       <S.PaddedItemThumbnail
@@ -99,33 +57,11 @@ const Items = () => {
         <TextsSection sectionTitle="유의사항" texts={itemCautionList} />
       )}
       {warning && <TextsSection sectionTitle="환불규정" texts={refundRuleList} />}
-      <S.HasActiveStateFullWidthButton
-        onClick={handlePurchaseBtnClick}
-        isActive={isActivePurchaseBtn}
-        className="purchase-btn"
-        backgroundColor={isActivePurchaseBtn ? 'sideMenuRed' : 'borderGray'}
-        color="white"
-      >
-        {isActiveOptionSelector || selectedOptionId !== -1 ? '구매하기' : '옵션 선택하기'}
-      </S.HasActiveStateFullWidthButton>
-      {isActiveOptionSelector && (
-        <>
-          <S.NotFullBackDrop onClick={() => handleSelector('close')} />
-          <OptionSelector
-            closeSelector={() => handleSelector('close')}
-            options={optionsWithDiscountRate}
-            handleSelect={handleOptionSelectorClick}
-          />
-        </>
-      )}
-      {selectedOptionId !== -1 && (
-        <SelectedOptionViewer
-          openSelector={() => {
-            handleSelector('open');
-          }}
-          option={getSelectedOptionText(selectedOptionId)}
-        />
-      )}
+      <OptionViewer
+        options={options}
+        minSellingPrice={minSellingPrice}
+        originalPrice={originalPrice}
+      />
     </S.ItemsLayout>
   );
 };
